+++ vec/src/Data/Vec/Pull.hs
@@ -28,8 +28,12 @@ module Data.Vec.Pull (
     _Cons,
     _head,
     _tail,
+    cons,
+    snoc,
     head,
     tail,
+    -- * Reverse
+    reverse,
     -- * Folds
     foldMap,
     foldMap1,
@@ -275,10 +279,19 @@ _tail :: I.Lens' (Vec ('S n) a) (Vec n a)
 _tail f (Vec v) = f (Vec (v . FS)) <&> \xs -> cons (v FZ) xs
 {-# INLINE _tail #-}
 
+-- | Cons an element in front of a 'Vec'.
 cons :: a -> Vec n a -> Vec ('S n) a
 cons x (Vec v) = Vec $ \i -> case i of
     FZ   -> x
     FS j -> v j
+
+-- | Add a single element at the end of a 'Vec'.
+--
+-- @since 0.2.1
+snoc :: forall a n. N.InlineInduction n => Vec n a -> a -> Vec ('S n) a
+snoc (Vec xs) x = Vec $ \i -> case F.isMax i of
+    Nothing -> x
+    Just i' -> xs i'
 
 -- | The first element of a 'Vec'.
 head :: Vec ('S n) a -> a
@@ -287,6 +300,20 @@ head (Vec v) = v FZ
 -- | The elements after the 'head' of a 'Vec'.
 tail :: Vec ('S n) a -> Vec n a
 tail (Vec v) = Vec (v . FS)
+
+-------------------------------------------------------------------------------
+-- Reverse
+-------------------------------------------------------------------------------
+
+-- | Reverse 'Vec'.
+--
+-- >>> reverse ('a' ::: 'b' ::: 'c' ::: VNil)
+-- 'c' ::: 'b' ::: 'a' ::: VNil
+--
+-- @since 0.2.1
+--
+reverse :: forall n a. N.InlineInduction n => Vec n a -> Vec n a
+reverse (Vec v) = Vec (v . F.mirror)
 
 -------------------------------------------------------------------------------
 -- Mapping