+++ fin/ChangeLog.md
@@ -1,5 +1,10 @@
 # Revision history for fin
 
+## 0.1.1
+
+- Add `isMin` and `isMax`
+- Add `mirror`, `weakenRight1` and `weakenLeft1`.
+
 ## 0.1
 
 - Rename `Fin` constructors to `FZ` and `FS`.